Add custom 'switch' operator
The switch statement has the following behavior:

(switch <tag-expr>
   ((<name1> <value1>)
    (<name2> <value2>)
    (else <value3>)))

If the result of evaluating tag-expr matches one
of the listed <name> values, return the specified
value, evaluated. Otherwise, if an else statement
is specified, return the value associated with the
else- if not, return '().

It is implemented as a derived statement, and is
transformed by the interpreter into a 'cond'.